+++ src/main/php/serializer/delegate/Tag.php
@@ -19,13 +19,13 @@ class Tag implements XmlSerializerDelegate
     /**
      * name of tag
      *
-     * @type  string
+     * @var  string
      */
     protected $tagName;
     /**
      * recurring element tag name for lists
      *
-     * @type  string
+     * @var  string
      */
     protected $elementTagName;
 
@@ -48,7 +48,7 @@ class Tag implements XmlSerializerDelegate
      * @param  \stubbles\xml\serializer\XmlSerializer  $xmlSerializer  serializer in case $value is not just a scalar value
      * @param  \stubbles\xml\XmlStreamWriter           $xmlWriter      xml writer to write serialized object into
      */
-    public function serialize($value, XmlSerializer $xmlSerializer, XmlStreamWriter $xmlWriter)
+    public function serialize($value, XmlSerializer $xmlSerializer, XmlStreamWriter $xmlWriter): void
     {
         $xmlSerializer->serialize($value, $xmlWriter, $this->tagName, $this->elementTagName);
     }